+++ src/TRUST_Modifs/Connex_components.h
@@ -20,9 +20,9 @@
 #include <TRUSTTabs_forward.h>
 #include <arch.h>
 
-long search_connex_components_local(const IntTab& elem_faces, const IntTab& faces_elem, IntVect& num_compo);
-long search_connex_components_local(const IntTab& elem_faces, const IntTab& faces_elem, DoubleTab& num_compo); // EB On sauvegarde num_compo et on le sort dans les latas. On a besoin d'un DoubleTab pour num_compo (pas de Champ_Fonc tel que .valeurs() renvoie un IntTab)
-long compute_graph_connex_components(const IntTab& graph, ArrOfInt& connex_components);
-long compute_global_connex_components(IntVect& num_compo, long nb_local_components);
-long compute_global_connex_components(DoubleTab& num_compo, long nb_local_components); // EB idem
+int search_connex_components_local(const IntTab& elem_faces, const IntTab& faces_elem, IntVect& num_compo);
+int search_connex_components_local(const IntTab& elem_faces, const IntTab& faces_elem, DoubleTab& num_compo); // EB On sauvegarde num_compo et on le sort dans les latas. On a besoin d'un DoubleTab pour num_compo (pas de Champ_Fonc tel que .valeurs() renvoie un IntTab)
+int compute_graph_connex_components(const IntTab& graph, ArrOfInt& connex_components);
+int compute_global_connex_components(IntVect& num_compo, int nb_local_components);
+int compute_global_connex_components(DoubleTab& num_compo, int nb_local_components); // EB idem
 #endif
